+++ envoy-control-runner/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/infrastructure/SynchronizationConfig.kt
@@ -1,7 +1,9 @@
 package pl.allegro.tech.servicemesh.envoycontrol.infrastructure
 
 import com.ecwid.consul.v1.ConsulClient
+import com.fasterxml.jackson.databind.ObjectMapper
 import io.micrometer.core.instrument.MeterRegistry
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -14,9 +16,11 @@ import pl.allegro.tech.servicemesh.envoycontrol.consul.ConsulProperties
 import pl.allegro.tech.servicemesh.envoycontrol.consul.synchronization.SimpleConsulInstanceFetcher
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.ControlPlaneClient
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.ControlPlaneInstanceFetcher
+import pl.allegro.tech.servicemesh.envoycontrol.synchronization.GzipRestTemplateControlPlaneClient
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RemoteClusterStateChanges
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RemoteServices
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RestTemplateControlPlaneClient
+import pl.allegro.tech.servicemesh.envoycontrol.utils.GzipUtils
 import java.lang.Integer.max
 import java.util.concurrent.Executors
 
@@ -35,12 +39,30 @@ class SynchronizationConfig {
     }
 
     @Bean
+    @ConditionalOnProperty(name = ["envoy-control.sync.gzip.enabled"], havingValue = "false", matchIfMissing = true)
     fun controlPlaneClient(restTemplate: RestTemplate, meterRegistry: MeterRegistry, remoteClusters: RemoteClusters) =
         RestTemplateControlPlaneClient(
             restTemplate = restTemplate,
             meterRegistry = meterRegistry,
             executors = Executors.newFixedThreadPool(max(remoteClusters.clusters.size, 1))
         )
+
+    @Bean
+    @Qualifier("controlPlaneClient")
+    @ConditionalOnProperty(name = ["envoy-control.sync.gzip.enabled"], havingValue = "true", matchIfMissing = false)
+    fun gZipControlPlaneClient(
+        restTemplate: RestTemplate,
+        meterRegistry: MeterRegistry,
+        gzipUtils: GzipUtils,
+        remoteClusters: RemoteClusters
+    ): GzipRestTemplateControlPlaneClient {
+        return GzipRestTemplateControlPlaneClient(
+            restTemplate = restTemplate,
+            meterRegistry = meterRegistry,
+            gzipUtils,
+            executors = Executors.newFixedThreadPool(max(remoteClusters.clusters.size, 1))
+        )
+    }
 
     @Bean
     fun remoteClusterStateChanges(
@@ -71,6 +93,11 @@ class SynchronizationConfig {
         ConsulClient(consulProperties.host, consulProperties.port),
         envoyControlProperties.sync.envoyControlAppName
     )
+
+    @Bean
+    fun gzipUtils(objectMapper: ObjectMapper): GzipUtils {
+        return GzipUtils(objectMapper)
+    }
 }
 
 data class RemoteClusters(val clusters: List<String>)